requirements.txt
fastapi
uvicorn
python-multipart
stripe
google-genai
python-jose[cryptography]
passlib[bcrypt]
boto3
bcrypt==3.2.2
